+++ src/main/java/id/perumdamts/mail/dto/core/folder/FolderCountDto.java
@@ -5,8 +5,8 @@ package id.perumdamts.mail.dto.core.folder;
  * Digunakan untuk menampilkan jumlah unread dan total mail.
  */
 public record FolderCountDto(
-        Integer folderId,
-        String folderName,
-        Long unread,
-        Long total
-) {}
+                Long folderId,
+                String folderName,
+                Long unread,
+                Long total) {
+}
