+++ src/main/java/id/perumdamts/mail/repository/core/jpa/MailRepository.java
@@ -3,5 +3,5 @@ package id.perumdamts.mail.repository.core.jpa;
 import id.perumdamts.mail.entity.core.Mail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MailRepository extends JpaRepository<Mail, Integer> {
+public interface MailRepository extends JpaRepository<Mail, Long> {
 }
